LayoutTests/fast/dom/shadow/distributed-pseudo-element-css-text.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
<style id="style1">
.foo::-webkit-distributed(div .bar::before) { display: block; }
</style>
</head>
<script>
description("Test for cssText of '::distributed()' rule.");
shouldBeEqualToString("document.getElementById('style1').sheet.cssRules.item(0).cssText", ".foo::-webkit-distributed(div .bar::before) { display: block; }");
</script>
<script src="../../js/resources/js-test-post.js"></script>
</html>
